feat: shows shibadex and polkadoge on top pairs
